requirements.txt
matplotlib==3.4.3
numpy==1.21.4
onnx==1.10.1
pandas==1.2.5
Pillow==8.4.0
pycuda==2021.1
tensorrt==8.0.3.4
torch==1.11.0a0+b6df043
torchvision==0.11.0a0
tqdm==4.62.3
